Detect/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Flask-Migrate
mysql-connector-python
newspaper3k
scikit-learn
joblib
passlib
python-dotenv
email_validator
# For PDF generation if needed on backend, jsPDF is for frontend
# reportlab 
# fpdf